@@ -1,76 +1,55 @@
 {
-  "name": "kamilrybacki",
-  "version": "1.0.0",
-  "private": true,
-  "description": "kamilrybacki",
-  "author": "kamil.rybacki",
-  "keywords": [
-    "gatsby"
-  ],
+  "name": "src",
+  "type": "module",
+  "version": "0.0.1",
   "scripts": {
-    "develop": "gatsby develop",
-    "start": "gatsby develop",
-    "build": "gatsby build",
-    "serve": "gatsby serve",
-    "clean": "gatsby clean",
-    "typecheck": "tsc --noEmit",
-    "deploy": "gatsby build && gh-pages -d public -b gh-pages"
+    "dev": "astro dev",
+    "start": "astro dev",
+    "build": "astro check && tsc --noEmit && astro build",
+    "preview": "astro preview",
+    "astro": "astro"
   },
   "dependencies": {
-    "@iconify/utils": "^1.0.32",
-    "@mdx-js/mdx": "^1.6.22",
-    "@mdx-js/react": "^1.6.22",
-    "babel-plugin-styled-components": "^2.0.7",
-    "emoji-dictionary": "^1.0.11",
-    "gatsby": "^4.12.1",
-    "gatsby-plugin-alias-imports": "^1.0.5",
-    "gatsby-plugin-graphql-codegen": "^3.1.1",
-    "gatsby-plugin-image": "^2.12.1",
-    "gatsby-plugin-mdx": "^3.13.0",
-    "gatsby-plugin-sharp": "^4.12.1",
-    "gatsby-plugin-styled-components": "^5.12.1",
-    "gatsby-source-filesystem": "^4.12.1",
-    "gatsby-transformer-sharp": "^4.12.1",
-    "gh-pages": "^3.2.3",
-    "iconify": "^1.4.0",
-    "node-fetch": "^2.6.7",
-    "parse-numeric-range": "^1.3.0",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-loading-skeleton": "^3.1.0",
-    "react-lorem-ipsum": "^1.4.10",
-    "react-markdown": "^8.0.3",
-    "react-modal": "^3.15.1",
-    "react-social-icons": "^5.13.0",
-    "react-syntax-highlighter": "^15.5.0",
-    "rehype-raw": "^6.1.1",
-    "remark-emoji": "^3.0.2",
-    "spinners-react": "^1.0.7",
-    "styled-components": "^5.3.5",
-    "tailwind-scrollbar": "^1.3.1",
-    "three": "^0.140.0",
-    "tsconfig-paths": "^4.0.0"
+    "@astrojs/check": "^0.2.0",
+    "@astrojs/mdx": "^1.0.0",
+    "@astrojs/react": "^3.0.0",
+    "@astrojs/tailwind": "^5.0.0",
+    "@astrojs/ts-plugin": "^1.1.3",
+    "@fontsource-variable/anybody": "^5.0.9",
+    "@fontsource-variable/raleway": "^5.0.9",
+    "@fontsource-variable/source-code-pro": "^5.0.9",
+    "@fontsource/black-and-white-picture": "^5.0.11",
+    "@fontsource/redacted-script": "^5.0.12",
+    "@fontsource/rubik-beastly": "^5.0.8",
+    "@fontsource/walter-turncoat": "^5.0.12",
+    "astro": "^3.0.8",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-svg": "^16.1.23",
+    "tailwind-styled-components": "^2.2.0",
+    "tailwindcss": "^3.3.3",
+    "typescript": "^5.2.2",
+    "node-html-parser": "^6.1.10"
   },
   "devDependencies": {
-    "@types/node": "^17.0.23",
-    "@types/react": "^17.0.44",
-    "@types/react-dom": "^17.0.15",
-    "@types/react-helmet": "^6.1.5",
-    "@types/react-syntax-highlighter": "^15.5.1",
-    "@types/styled-components": "^5.1.25",
-    "@types/tailwindcss": "^3.0.10",
-    "@typescript-eslint/eslint-plugin": "^5.25.0",
-    "@typescript-eslint/parser": "^5.25.0",
-    "autoprefixer": "^10.4.5",
-    "eslint": "^8.15.0",
+    "@types/react": "^18.2.21",
+    "@types/react-dom": "^18.2.7",
+    "@typescript-eslint/parser": "^6.7.2",
+    "eslint": "^8.50.0",
     "eslint-config-google": "^0.14.0",
-    "eslint-plugin-react": "^7.30.0",
-    "gatsby-plugin-postcss": "^5.12.1",
-    "gatsby-plugin-react-helmet": "^5.15.0",
-    "postcss": "^8.4.12",
-    "react-helmet": "^6.1.0",
-    "tailwind-styled-components": "^2.1.7",
-    "tailwindcss": "^3.0.24",
-    "typescript": "^3.8.3"
+    "eslint-plugin-astro": "^0.29.0",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-config-airbnb-typescript": "^17.1.0",
+    "eslint-config-prettier": "^9.0.0",
+    "eslint-plugin-import": "^2.28.1",
+    "eslint-plugin-prettier": "^5.0.0",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-simple-import-sort": "^10.0.0",
+    "eslint-plugin-tailwindcss": "^3.13.0",
+    "eslint-plugin-unused-imports": "^3.0.0",
+    "prettier": "^3.0.3",
+    "prettier-plugin-astro": "^0.12.0"
   }
 }
