@@ -1,13 +1,20 @@
-const customTheme = require('./src/style/theme');
-
-module.exports = {
-  content: ['./src/**/*.{js,ts,jsx,tsx}'],
-  theme: customTheme,
-  plugins: [
-    function({addVariant}) {
-      addVariant('child', '& > *');
-      addVariant('child-hover', '& > *:hover');
+/** @type {import('tailwindcss').Config} */
+export const content = ["./src/**/*.{astro,html,js,jsx,md,mdx,svelte,ts,tsx,vue}"];
+export const theme = {
+    colors: {
+        foreground: "#C4D8ED",
+        background: "#090E11",
+        accent: "#E5F2FF",
+        glow: "#C4D8ED",
     },
-    require('tailwind-scrollbar'),
-  ],
+    extend: {
+        fontFamily: {
+            display: ["Anybody Variable"],
+            body: ["Raleway Variable"],
+            code: ["Source Code Pro Variable"],
+            handwriting: ["Walter Turncoat"],
+            abstract: ["Redacted Script"],
+        },
+    },
 };
+export const plugins = [];
